Added Api keys for all the apis

+++ Route/Employer.js
@@ -9,18 +9,22 @@ router.route("/").get(async (req, res) => {
 });
 
 router.route("/:id").get(async (req, res) => {
-  await employer
-    .findById(req.params.id)
-    .then((emp) => {
-      if (!emp) {
-        return res
-          .status(404)
-          .json("User not Find with ID:" + emp)
-          .send();
-      }
-      res.send(emp);
-    })
-    .catch((err) => res.status(400).json("Error" + err));
+  if (req.query.api_key == apiKey) {
+    await employer
+      .findById(req.params.id)
+      .then((emp) => {
+        if (!emp) {
+          return res
+            .status(404)
+            .json("User not Find with ID:" + emp)
+            .send();
+        }
+        res.send(emp);
+      })
+      .catch((err) => res.status(400).json("Error" + err));
+  } else {
+    res.send("unathorized access");
+  }
 });
 
 router.route("/addemployer").post(async (req, res) => {
@@ -47,34 +51,42 @@ router.route("/addemployer").post(async (req, res) => {
 });
 
 router.route("/delete/:id").delete(async (req, res) => {
-  const id = req.params.id;
-  await employer
-    .findByIdAndDelete(id)
-    .then((emp) => {
-      if (!emp) {
-        return res.status(404).json("employer not Find with ID").send();
-      }
-      res.send(emp);
-    })
-    .catch((error) => {
-      res.status(500).send(error);
-    });
+  if (req.query.api_key == apiKey) {
+    const id = req.params.id;
+    await employer
+      .findByIdAndDelete(id)
+      .then((emp) => {
+        if (!emp) {
+          return res.status(404).json("employer not Find with ID").send();
+        }
+        res.send(emp);
+      })
+      .catch((error) => {
+        res.status(500).send(error);
+      });
+  } else {
+    res.send("unathorized access");
+  }
 });
 
 router.route("/update/:id").put(async (req, res) => {
-  const id = req.params.id;
+  if (req.query.api_key == apiKey) {
+    const id = req.params.id;
 
-  await employer
-    .findByIdAndUpdate(id, req.body, { new: true })
-    .then((emp) => {
-      if (!emp) {
-        return res.status(404).json("employer not Find with ID").send();
-      }
-      res.send(emp);
-    })
-    .catch((error) => {
-      res.status(500).send(error);
-    });
+    await employer
+      .findByIdAndUpdate(id, req.body, { new: true })
+      .then((emp) => {
+        if (!emp) {
+          return res.status(404).json("employer not Find with ID").send();
+        }
+        res.send(emp);
+      })
+      .catch((error) => {
+        res.status(500).send(error);
+      });
+  } else {
+    res.send("unathorized access");
+  }
 });
 
 module.exports = router;
